added auto-include module
