+++ src/components/play-activity-page/play-activity-page.js
@@ -1,6 +1,7 @@
 import ko from 'knockout';
 import templateMarkup from 'text!./play-activity-page.html';
 import Preferences from 'app/preferences';
+import AutoIncluder from 'components/editor/auto-include'
 import * as SysGlobalObservables from 'app/sys-global-observables';
 
 class PlayActivityPage {
@@ -15,57 +16,7 @@ class PlayActivityPage {
         } else {
             this.setParamsFromDefaults();
         }
-        var self = this;
-        this.createMapping();
-    }
-
-    createMapping() {
-        var includeMap = {};
-        $.getJSON("https://cs-education.github.io/sysassets/man_pages/syscall_metadata.min.json", function(data) {
-            data.forEach(function(element) {
-                element.functions.forEach(function(func) {
-                    includeMap[func.name] = [];
-                    element.defines.forEach(function(define) {
-                        includeMap[func.name].push("#define " + define.text);
-                    });
-                    element.includes.forEach(function(include) {
-                        includeMap[func.name].push("#include <" + include.file_path + ">");
-                    });
-                });
-            });
-            self.IncludeMap = includeMap;
-        });
-    }
-
-    addMissingHeaders() {
-        var getter = this.editorParams.editorTextGetter();
-        var text = getter().split('\n');
-        var currentHeaders = text.filter(function(element) {
-            return element[0] === '#';
-        });
-        var editor = SysGlobalObservables.Editor.aceEditor;
-        var session = editor.session;
-        var length = session.getLength();
-        var includeMap = self.IncludeMap;
-        var headers = [];
-        for (var row = 1; row <= length; row++) {
-            var tokens = session.getTokens(row);
-            tokens.forEach(function(token) {
-                var syntax = token.value;
-                if (syntax in includeMap) {
-                    includeMap[syntax].forEach(function(header) {
-                        if (headers.indexOf(header) == -1 && currentHeaders.indexOf(header) == -1) {
-                            headers.push(header);
-                        }
-                    });
-                }
-            });
-        }
-        headers.forEach(function(header) {
-            text.unshift(header);
-        });
-        var final = text.join('\n');
-        SysGlobalObservables.Editor.setAceText(final);
+        this.autoIncluder = new AutoIncluder();
     }
 
     setParamsFromActivity(playActivity) {
@@ -166,7 +117,7 @@ class PlayActivityPage {
         this.editorParams.editorTextGetter = ko.observable(() => '');
 
         var compile = () => {
-            this.addMissingHeaders();
+            this.autoIncluder.addMissingHeaders(this);
             var buildCmd = this.compilerParams.buildCmd();
             (SysGlobalObservables.runCode())(buildCmd);
         };
